network rendering and licensing stuffs
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@842 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/stringutils.cpp
@@ -850,7 +850,7 @@ const std::string forceCopyString(const std::string& s)
 }
 
 
-const std::string getNiceByteSize(size_t x)
+const std::string getNiceByteSize(uint64 x)
 {
 	assert(x >= 0);
 	if(x < 1024)
